fixed null exception bug

+++ resources/views/templates/modules/comment.php
@@ -6,9 +6,9 @@
         <mark class="info"><?php echo $comment_author->username; ?></mark>
         <p><?php echo $comment_body ?></p>
         <?php
-            if($comment_author->id == $p->id){
-                echo "<a href=\"/crash/athenaeum/comment/delete?id_comment=$comment->id_comment&uri_redirect_back=$publication->uri\"><mark class=\"danger\">[delete]</mark></a>"; 
-            }
+        if(isset($p)){
+            if($comment_author->id == $p->id) echo "<a href=\"/crash/athenaeum/comment/delete?id_comment=$comment->id_comment&uri_redirect_back=$publication->uri\"><mark class=\"danger\">[delete]</mark></a>"; 
+        }
         ?>
     </div>
 </div>
